% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{auroc}
\alias{auroc}
\title{Compute the area under an ROC curve.}
\usage{
auroc(x, ...)
}
\arguments{
\item{x}{An appropriate object.}

\item{...}{Further arguments passed to method.}
}
\value{
The area under the ROC curve(s).
}
\description{
Generic function for computing the area under an ROC curve.
}
\seealso{
\code{\link{auroc.ep}}, \code{\link{auroc.ep.list}}
}
