% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{bf.bnpostmcmc}
\alias{bf.bnpostmcmc}
\title{Bayes Factors.}
\usage{
\method{bf}{bnpostmcmc}(bn1, bn2, data, ...)
}
\arguments{
\item{bn1}{A bn}

\item{bn2}{...}

\item{data}{...}

\item{...}{further arguments}
}
\description{
method description
}
\seealso{
\code{\link{bf}}
}
