% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{bf}
\alias{bf}
\title{Bayes Factor.}
\usage{
bf(bn1, bn2, data, ...)
}
\arguments{
\item{bn1}{A bn}

\item{bn2}{A bn}

\item{data}{data}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{bf.bnpostmcmc}}
}
