\name{bnpostmcmc.list}
\alias{bnpostmcmc.list}
\title{List of BN Posteriors from MCMC.}
\usage{
  \method{bnpostmcmc}{list} (...)
}
\arguments{
  \item{...}{Further arguments passed to method}
}
\description{
  method description
}
\seealso{
  \code{\link{bnpostmcmc}},
  \code{\link{gp.bnpostmcmc.list}},
  \code{\link{ep.bnpostmcmc.list}}
}

