\name{BNSamplerGrzeg}
\alias{BNSamplerGrzeg}
\title{method name...}
\usage{BNSamplerGrzeg(data, initial, prior, return="network",
    logScoreFUN=defaultLogScoreFUN(),
    logScoreParameters=list(hyperparameters = "qi"), constraint,
    modejumping=F, verbose=F, keepTape=F,
    grzeg=list(grzegMoveProbability = 0.5))}
\description{method name}
\details{method description}
\arguments{\item{data}{...}
\item{initial}{...}
\item{prior}{...}
\item{return}{...}
\item{logScoreFUN}{...}
\item{logScoreParameters}{...}
\item{constraint}{...}
\item{modejumping}{NOT IMPLEMENTED}
\item{verbose}{...}
\item{keepTape}{...}
\item{grzeg}{...}}
