\name{burnin}
\alias{burnin}
\title{Extract burnin}
\usage{
  burnin(x, ...)

  burnin(x, ...)
}
\arguments{
  \item{x}{A sampler}

  \item{...}{Further arguments passed to method}

  \item{x}{A sampler}

  \item{...}{Further arguments passed to method}
}
\value{
  An integer amount of burnin

  An integer amount of burnin
}
\description{
  Extracts the amount of burnin used in a sampler

  Extracts the amount of burnin used in a sampler
}
\seealso{
  \code{\link{steps.sampler}}

  \code{\link{steps.sampler}}
}

