% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{burnin.sampler}
\alias{burnin.sampler}
\title{Extract burnin}
\usage{
\method{burnin}{sampler}(x, ...)
}
\arguments{
\item{x}{A sampler}

\item{...}{Further arguments, currently unused}
}
\value{
An integer amount of burnin
}
\description{
Extracts the amount of burnin used in a sampler
}
