% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-exact.R
\name{conductance}
\alias{conductance}
\title{Conductance.}
\usage{
conductance(x, states)
}
\arguments{
\item{x}{...}

\item{states}{...}
}
\description{
method description
}
\seealso{
\code{\link{eht}}, \code{\link{hp}}, \code{\link{tp}}
}
