\name{cumep.bnpostmcmc.list}
\alias{cumep.bnpostmcmc.list}
\title{Cumulative edge probabilities.}
\usage{
  \method{cumep}{bnpostmcmc.list} (x, ...)
}
\arguments{
  \item{x}{...}

  \item{...}{Further arguments passed to method}
}
\description{
  method description
}
\seealso{
  \code{\link{cumep}}
}

