% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{cumep.list}
\alias{cumep.list}
\title{Cumulative edge probabilities.}
\usage{
\method{cumep}{list}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{cumep}}
}
