% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{cumep}
\alias{cumep}
\title{Cumulative edge probabilities.}
\usage{
cumep(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{cumep.bnpostmcmc.list}}
}
