\name{cummean.matrix}
\alias{cummean.matrix}
\title{Compute the cumulative means of the columns of a matrix x.}
\usage{\method{cummean}{matrix} (x, ...)}
\description{Compute the cumulative means of the columns of a matrix x.
ie each column is treated separately}
\value{A matrix of the same dimension as x, with the cumulative values.}
\arguments{\item{x}{A matrix}
\item{...}{Further arguments (unused)}}
