% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{cummean}
\alias{cummean}
\title{Cumulative mean.}
\usage{
cummean(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{cummean.matrix}}
}
