% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{cumtvd.list}
\alias{cumtvd.list}
\title{List of cumulative total variation distance.}
\usage{
cumtvd.list(...)
}
\arguments{
\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{cumtvd}}
}
