\name{dotplot.ep.list}
\alias{dotplot.ep.list}
\title{method name...}
\usage{\method{dotplot}{ep.list} (eplist, subsetBy="Exact", head=30, ...)}
\description{method name}
\details{method description}
\arguments{\item{eplist}{...}
\item{subsetBy}{...}
\item{head}{...}
\item{...}{Further arguments passed to method}}
