\name{dotplot.gp}
\alias{dotplot.gp}
\title{Dotplot of posterior graph probabilities.}
\usage{\method{dotplot}{gp} (gp, head=30, ...)}
\description{Dotplot of posterior graph probabilities.}
\details{method description}
\seealso{\code{\link{xyplot.gp}}}
\arguments{\item{gp}{...}
\item{head}{...}
\item{...}{Further arguments passed to method}}
