% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-pt.R
\name{drawPT}
\alias{drawPT}
\title{Undocumented.}
\usage{
drawPT(
  samplers,
  data,
  n,
  temperatures,
  pswap = 0.2,
  burnin = 0,
  verbose = T,
  currentNetwork
)
}
\arguments{
\item{samplers}{...}

\item{data}{...}

\item{n}{...}

\item{temperatures}{...}

\item{pswap}{...}

\item{burnin}{...}

\item{verbose}{...}

\item{currentNetwork}{BUG found by R-check}
}
\description{
method description
}
\seealso{
\code{\link{BNSamplerPT}}, \code{\link{draw}}
}
