% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-exact.R
\name{eht}
\alias{eht}
\title{Expected hitting time.}
\usage{
eht(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{eht.bnpost}}
}
