\name{enforceForbiddenConstraint}
\alias{enforceForbiddenConstraint}
\title{Undocumented.}
\usage{
  enforceForbiddenConstraint(x, constraint)
}
\arguments{
  \item{x}{...}

  \item{constraint}{...}
}
\description{
  method description
}
\seealso{
  \code{\link{satisfiesConstraint}},
  \code{\link{setupConstraint}}
}

