\name{entropy.bnpost}
\alias{entropy.bnpost}
\title{Entropy.}
\usage{\method{entropy}{bnpost} (x, logNetworkPriors, ...)}
\description{Entropy.}
\details{method description}
\seealso{\code{\link{entropy}}}
\arguments{\item{x}{...}
\item{logNetworkPriors}{...}
\item{...}{Further arguments (unused)}}
