% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{entropy}
\alias{entropy}
\title{Entropy.}
\usage{
entropy(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{entropy.bnpost}}
}
