\name{ep.bnpost}
\alias{ep.bnpost}
\title{Posterior edge probabilities.}
\usage{\method{ep}{bnpost} (x, ...)}
\description{Posterior edge probabilities.}
\details{method description}
\seealso{\code{\link{ep}}, \code{\link{gp.bnpost}}}
\arguments{\item{x}{...}
\item{...}{Further arguments (unused)}}
