\name{ep}
\alias{ep}
\title{Posterior edge probabilities.}
\usage{ep(x, ...)}
\description{Posterior edge probabilities.}
\details{method description}
\seealso{\code{\link{ep.bnpostmcmc.list}}, \code{\link{ep.parental.list}}, 
\code{\link{ep.bnpost}}, \code{\link{ep.table}},
\code{\link{ep.parental.contingency}}}
\arguments{\item{x}{...}
\item{...}{Further arguments passed to method}}
