\name{epmx}
\alias{epmx}
\title{Edge probabilties matrix.}
\usage{
  epmx(x, ...)
}
\arguments{
  \item{x}{...}

  \item{...}{Further arguments passed to method}
}
\description{
  method description
}
\seealso{
  \code{\link{epmx.bnpostmcmc.list}},
  \code{\link{splom.epmx}},
  \code{\link{splom.bnpostmcmc.list}}
}

