\name{epmxFromET}
\alias{epmxFromET}
\title{Retreive edge prob matrix from edge totals.}
\usage{epmxFromET(y)}

\description{
  Retreive edge prob matrix from edge totals.
}

\details{
  Returns matrix for epmx from the edge totals matrix
  collected online.
}
\value{An epmx}
\arguments{
  \item{y}{An \code{bnpostmcmc}}
}
