% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{getBannedFromConstraint}
\alias{getBannedFromConstraint}
\title{Resample a pair of nodes together.}
\usage{
getBannedFromConstraint(constraint)
}
\arguments{
\item{constraint}{..}
}
\description{
Resample a pair of nodes together.
}
