% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{getRequiredFromConstraint}
\alias{getRequiredFromConstraint}
\title{Resample a pair of nodes together.}
\usage{
getRequiredFromConstraint(constraint)
}
\arguments{
\item{constraint}{..}
}
\description{
Resample a pair of nodes together.
}
