\name{gp.bnpost}
\alias{gp.bnpost}
\title{method name...}
\usage{\method{gp}{bnpost} (x, logNetworkPriors, log=F, pretty=F, ...)}
\description{method name}
\details{method description}
\arguments{\item{x}{...}
\item{logNetworkPriors}{...}
\item{log}{...}
\item{pretty}{...}
\item{...}{Further arguments (unused)}}
