% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-mcmc.R
\name{gp.bnpostmcmc.list}
\alias{gp.bnpostmcmc.list}
\title{Posterior graph probabilities.}
\usage{
\method{gp}{bnpostmcmc.list}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{gp}}, \code{\link{gp.bnpostmcmc}}
}
