% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{gp}
\alias{gp}
\title{Posterior graph probabilities.}
\usage{
gp(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{summary.gp}},
  \code{\link{gp.bnpostmcmc}}, \code{\link{gp.bnpost}},
  \code{\link{gp.bnpostmcmc.list}}, \code{\link{gp.list}}
}
