\name{gp.sampler}
\alias{gp.sampler}
\title{Retreive graph probabilities from sampler}
\usage{
  \method{gp}{sampler} (x, head = 10000, ...)
}
\arguments{
  \item{x}{A sampler}

  \item{head}{The number of graphs to consider}

  \item{...}{Further arguments, currently unused}
}
\description{
  \code{gp} method for class "\code{sampler}"
}

