% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-exact.R
\name{hp}
\alias{hp}
\title{Hitting probability.}
\usage{
hp(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{eht}}, \code{\link{conductance}}, \code{\link{tp}}
}
