\name{intAsFDF}
\alias{intAsFDF}
\title{Convert a data frame of integers to factors.}
\usage{
  intAsFDF(x)
}
\arguments{
  \item{x}{an object of class \code{data.frame}}
}
\value{
  The data.frame with columns converted to factors
}
\description{
  Converts a \code{data.frame} that consists columns of
  integers into a \code{data.frame} consisting of factos
}
\seealso{
  \code{\link{fdfAsInt}}
}

