% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structmcmc.R
\name{internal-onload}
\alias{internal-onload}
\alias{.onLoad}
\title{Internal function.}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{...}

\item{pkgname}{...}
}
\description{
Internal function.
}
