\name{is.valid.hyp}
\alias{is.valid.hyp}
\title{Checks whether the supplied hyperparameters parameter is valid.}
\usage{\method{is.valid}{hyp} (x)}
\description{Checks whether the supplied hyperparameters parameter is valid.}
\value{A logical of length 1.}
\arguments{\item{x}{The hyperparameters to test.}}
