% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{kronecker_delta}
\alias{kronecker_delta}
\title{Metric: Kronecker delta}
\usage{
kronecker_delta(predictions, true)
}
\arguments{
\item{predictions}{A factor variable}

\item{true}{A factor variable}
}
\description{
Metric: Kronecker delta
}
