\name{length.bnpostmcmc}
\alias{length.bnpostmcmc}
\title{Number of samples drawn.}
\usage{
  \method{length}{bnpostmcmc} (x, ...)
}
\arguments{
  \item{x}{An object of class 'bnpostmcmc'}

  \item{...}{Further arguments passed to method}
}
\value{
  The number of samples in x
}
\description{
  Returns the number of samples draw in the supplied object
  of class 'bnpostmcmc' x.
}
\seealso{
  \code{\link{bnpostmcmc}}.
}

