\name{levelplot.bnpostmcmc}
\alias{levelplot.bnpostmcmc}
\title{Levelplot of BN Posterior from MCMC.}
\usage{\method{levelplot}{bnpostmcmc} (x)}

\description{
  Levelplot of BN Posterior from MCMC.
}

\details{
  method description
}
\seealso{\code{\link{plot.bnpostmcmc}}}
\arguments{
  \item{x}{...}
}
