\name{levelplot.ep}
\alias{levelplot.ep}
\title{Levelplot of posterior edge probabilities.}
\usage{\method{levelplot}{ep} (ep)}
\description{Levelplot of posterior edge probabilities.}
\details{method description}
\seealso{\code{\link{levelplot.ep.list}}, \code{\link{dotplot.ep}}}
\arguments{\item{ep}{...}}
