\name{localPartitionFunction}
\alias{localPartitionFunction}
\title{method name...}
\usage{localPartitionFunction(currentNetwork, proposalNetwork, node,
    required=integer(0), logScoreOnlineFUN, logScoreParameters, cache)}
\description{method name}
\details{method description}
\arguments{\item{currentNetwork}{...}
\item{proposalNetwork}{...}
\item{node}{...}
\item{required}{...}
\item{logScoreOnlineFUN}{...}
\item{logScoreParameters}{...}
\item{cache}{...}}
