\name{logScoreMultDir}
\alias{logScoreMultDir}
\title{Multinomial-Dirichlet Log marginal likelihood.}
\usage{logScoreMultDir(x, ...)}
\description{Multinomial-Dirichlet Log marginal likelihood.}
\details{method description}
\seealso{\code{\link{logScoreMultDir.bn}},
\code{\link{logScoreMultDir.bn.list}},
\code{\link{logScoreMultDirOffline}},
\code{\link{logScoreMultDirIncremental}}}
\arguments{\item{x}{A \code{bn}.}
\item{...}{Further arguments, passed to method}}
