% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-normal.R
\name{logScoreNormalPrepare}
\alias{logScoreNormalPrepare}
\title{Internal functions.}
\usage{
logScoreNormalPrepare(data, logScoreParameters, checkInput = T)
}
\arguments{
\item{data}{A data.frame, with columns being factors giving the values of
each random variable.}

\item{logScoreParameters}{A list with the following components:
\describe{
  \item{data}{A matrix with columns giving the values of each random
              variable.}
  \item{nl}{A numeric vector of length nNodes(currentBN), specifying the
            number of levels that each random variable takes.}
}}

\item{checkInput}{A logical of length 1, specifying whether to check the
inputs to the function.}
}
\value{
A list with the contents of logScoreParameters, with the following
  components added or altered:
  \describe{
    \item{data}{A matrix with columns giving the value of each random
                variable.}
    \item{nl}{A numeric vector of length ncol(data), specifying the number
              of levels that each random variable takes.}
  }
}
\description{
Convert a data frame to the appropriate format for the fast/incremental
logScoreNormal functions, and return as part of the logScoreParameters
list.
}
\details{
In particular, the data is converted to a matrix, and the factor levels
taken integer values from 0, 1, .... ie not on 1, 2, 3.
}
\seealso{
\code{\link{logScoreNormal}}
}
