% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{makeRowPredictions}
\alias{makeRowPredictions}
\title{Make row predictions}
\usage{
makeRowPredictions(
  row,
  predictors,
  param,
  type = c("response", "probabilities")
)
}
\arguments{
\item{row}{A vector containing the predictors in the model}

\item{predictors}{The parents of the node}

\item{param}{A list of parameters}

\item{type}{The type of prediction required.}
}
\description{
Make row predictions
}
