% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{map}
\alias{map}
\title{Maximum aposteriori graph.}
\usage{
map(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{map.bnpost}}, \code{\link{map.bnpostmcmc}}
}
