% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{map.sampler}
\alias{map.sampler}
\title{Retreive MAP from sampler}
\usage{
\method{map}{sampler}(x, ...)
}
\arguments{
\item{x}{A sampler}

\item{...}{Further arguments, currently unused}
}
\description{
\code{map} method for class "\code{sampler}"
}
