\name{mwep.bnpostmcmc.list}
\alias{mwep.bnpostmcmc.list}
\title{Moving window edge probabilities.}
\usage{
  \method{mwep}{bnpostmcmc.list} (x, window = 10, ...)
}
\arguments{
  \item{x}{...}

  \item{window}{...}

  \item{...}{Further arguments passed to method}
}
\description{
  method description
}
\seealso{
  \code{\link{mwep}}
}

