% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mwep.list}
\alias{mwep.list}
\title{Moving window edge probabilities.}
\usage{
\method{mwep}{list}(x, window = 10, ...)
}
\arguments{
\item{x}{...}

\item{window}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{mwep}}
}
