% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mwep}
\alias{mwep}
\title{Moving window edge probiilities.}
\usage{
mwep(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{mwep.bnpostmcmc.list}}
}
