% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mwmean}
\alias{mwmean}
\title{Moving window mean.}
\usage{
mwmean(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{mwmean.matrix}}
}
