% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.rjbg.R
\name{new_queue}
\alias{new_queue}
\title{Priority queue}
\usage{
new_queue(threshold)
}
\arguments{
\item{threshold}{The amount below the top that should be kept}
}
\description{
A priority queue, at least of sorts
}
\details{
Based upon Hadley Wickham's memoise (MIT License)
http://cran.r-project.org/web/packages/memoise/
}
