% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.rjbg.R
\name{new_stack}
\alias{new_stack}
\title{Most-recently used stack.}
\usage{
new_stack(size = 1000L)
}
\arguments{
\item{size}{An integer. The stack size}
}
\description{
Stack that is aware of which items have been used most recently.
}
