\name{nonDescendants}
\alias{nonDescendants}
\title{Find nonDescendants.}
\usage{
  nonDescendants(x, node)
}
\arguments{
  \item{x}{A routes matrix}

  \item{node}{A node}
}
\description{
  Find the non-descendants of a node
}
\seealso{
  \code{\link{routesAddEdges}},
  \code{\link{routesRemoveEdges}}
}

