% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{print.sampler}
\alias{print.sampler}
\title{Print a sampler}
\usage{
\method{print}{sampler}(x, ...)
}
\arguments{
\item{x}{A sampler}

\item{...}{Further arguments, currently unused}
}
\description{
Prints a sampler
}
