\name{requireSomethingFromEachParent}
\alias{requireSomethingFromEachParent}
\title{Require something from each parent.}
\usage{requireSomethingFromEachParent(numberOfNodes, nodesSeq,
    possibleParents, change, descendantsList)}
\description{Require something from each parent.}
\value{A list}
\arguments{\item{numberOfNodes}{The number of nodes}
\item{nodesSeq}{A sequence \code{1:numberOfNodes}}
\item{possibleParents}{The output of \code{\link{getPossibleParents}}}
\item{change}{A numeric vector, containing the nodes whose parents
are to be changed.}
\item{descendantsList}{The output of \code{\link{descendantsList}}}}
