\name{residsNormal}
\alias{residsNormal}
\title{Residuals from a Normal model}
\usage{
  residsNormal(x, train, test, metric = square_error)
}
\arguments{
  \item{x}{A BN}

  \item{train}{A data frame of training data}

  \item{test}{A data frame of test data}

  \item{metric}{A function that measures the distance
  between the predictions and the true values}
}
\description{
  ...
}

