% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{residsNormalNode}
\alias{residsNormalNode}
\title{Residuals for a single node for a Normal model}
\usage{
residsNormalNode(node, x, train, test, metric, verbose)
}
\arguments{
\item{node}{An integer, giving the node}

\item{x}{A BN}

\item{train}{A data frame of training data}

\item{test}{A data frame of test data}

\item{metric}{A function that measures the distance between the
predictions and the true values}
}
\description{
Residuals for a single node for a Normal model
}
