\name{residualsMultDir}
\alias{residualsMultDir}
\title{Residuals from a Multinomial-Dirichlet model}
\usage{
  residualsMultDir(x, train, test, metric =
  kronecker_delta, verbose = F)
}
\arguments{
  \item{x}{A BN}

  \item{train}{A data frame of training data}

  \item{test}{A data frame of test data}

  \item{metric}{A function that measures the distance
  between the predictions and the true values}

  \item{verbose}{Logical indicating whether verbose output
  should be given}
}
\description{
  ..
}

