\name{residualsMultDirNode}
\alias{residualsMultDirNode}
\title{Residuals for a single node for a Multinomial-Dirichlet model}
\usage{
  residualsMultDirNode(node, x, param, test, metric,
  verbose)
}
\arguments{
  \item{node}{An integer, giving the node}

  \item{x}{A BN}

  \item{test}{A data frame of test data}

  \item{metric}{A function that measures the distance
  between the predictions and the true values}

  \item{verbose}{Logical indicating whether verbose output
  should be given}
}
\description{
  Residuals for a single node for a Multinomial-Dirichlet
  model
}

