\name{rowSums2}
\alias{rowSums2}
\title{Fast, dangerous row sums.}
\usage{
  rowSums2(x)
}
\arguments{
  \item{x}{A matrix.}
}
\value{
  A vector of row sums.
}
\description{
  A fast, simple version of \code{rowSums}. This version
  only handles matrices.
}
\details{
  Note that no sanity checks are performed on the input.
}

