\name{satisfiesConstraint}
\alias{satisfiesConstraint}
\title{Undocumented.}
\usage{satisfiesConstraint(x, constraint)}
\description{Undocumented.}
\details{method description}
\seealso{\code{\link{setupConstraint}},
\code{\link{enforceForbiddenConstraint}}}
\arguments{\item{x}{...}
\item{constraint}{...}}
