\name{scoreBIC}
\alias{scoreBIC}
\title{method name...}
\usage{scoreBIC(bn, data, nodes=seq_along(bn), mlcache=new.env(hash = T),
    nijkcache=new.env(hash = T))}
\description{method name}
\details{method description}
\arguments{\item{bn}{...}
\item{data}{...}
\item{nodes}{...}
\item{mlcache}{...}
\item{nijkcache}{...
#'}}
