% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{setupConstraint}
\alias{setupConstraint}
\title{Setup constraint.}
\usage{
setupConstraint(constraint, initial)
}
\arguments{
\item{constraint}{..}

\item{initial}{..}
}
\description{
...
}
\seealso{
\code{\link{satisfiesConstraint}},
  \code{\link{enforceForbiddenConstraint}}
}
