\name{splom.bnpostmcmc.list}
\alias{splom.bnpostmcmc.list}
\title{Scatterplot matrix of edge probabilities between runs.}
\usage{
  \method{splom}{bnpostmcmc.list} (x, start = 1, end,
    plot = F)
}
\arguments{
  \item{x}{...}

  \item{start}{...}

  \item{end}{...}

  \item{plot}{...}
}
\description{
  method description
}
\seealso{
  \code{\link{epmx}}, \code{\link{splom.epmx}},
  \code{\link{splom}}
}

