% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{square_error}
\alias{square_error}
\title{Metric: Square error}
\usage{
square_error(predictions, true)
}
\arguments{
\item{predictions}{A numeric variable}

\item{true}{A numeric variable}
}
\description{
Metric: Square error
}
