\name{statistics}
\alias{statistics}
\title{Extract statistics from a sampler.}
\usage{statistics(x, ...)}

\description{
  Extract statistics from a sampler.
}

\details{
  Extracts the statistics collected during an MCMC run
}
\value{A list of statistics collected during the MCMC run.}
\seealso{\code{\link{statistics.sampler}}}
\arguments{
  \item{x}{A sampler}
  \item{...}{Further arguments passed to method}
}
