\name{steps}
\alias{steps}
\title{Number of samples drawn.}
\usage{
  steps(x, ...)
}
\arguments{
  \item{x}{A sampler}

  \item{...}{Further arguments passed to method}
}
\value{
  The number of samples (steps) that have been drawn.
}
\description{
  Returns the number of samples (MCMC steps) drawn in the
  supplied sampler.
}
\seealso{
  \code{\link{steps.sampler}}
}

