\name{summary.gp}
\alias{summary.gp}
\title{Summary of posterior graph probabilities.}
\usage{\method{summary}{gp} (object, ...)}
\description{Summary of posterior graph probabilities.}
\details{method description}
\seealso{\code{\link{gp}}}
\arguments{\item{object}{...}
\item{...}{Further arguments passed to method}}
