% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{summary.sampler}
\alias{summary.sampler}
\title{Summarising MCMC samplers}
\usage{
\method{summary}{sampler}(x, ...)
}
\arguments{
\item{x}{A sampler}

\item{...}{Further arguments, currently unused}
}
\description{
\code{summary} method for class "\code{sampler}"
}
