% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{top}
\alias{top}
\title{Top posterior graph.}
\usage{
top(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{top.bnpost}}, \code{\link{top.bnpostmcmc}}
}
