% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-exact.R
\name{tp}
\alias{tp}
\title{Transition probabilities.}
\usage{
tp(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{eht}}, \code{\link{hp}}, \code{\link{conductance}}
}
