\name{xyplot.cumtvd.list}
\alias{xyplot.cumtvd.list}
\title{Plot of cumulative total variation distance.}
\usage{\method{xyplot}{cumtvd.list} (cumtvdlist)}

\description{
  Plot of cumulative total variation distance.
}

\details{
  method description
}
\seealso{\code{\link{xyplot.cumtvd.list}}}
\arguments{
  \item{cumtvdlist}{...}
}
