\name{xyplot.epmx}
\alias{xyplot.epmx}
\title{Plot of cumulative edge probabilities.}
\usage{\method{xyplot}{epmx} (x, subset)}

\description{
  Plot of cumulative edge probabilities.
}

\details{
  Returns a xyplot of the cumulative edge probabilities
  through time for bnpostmcmc.list and bvspostmcmc.list
}
\value{...}
\seealso{\code{\link{epmx}}}
\arguments{
  \item{x}{...}
  \item{subset}{A numeric vector. This specifies the nodes between which
the epmx should be displayed for.}
}
