\name{xyplot.gp.list}
\alias{xyplot.gp.list}
\title{method name...}
\usage{\method{xyplot}{gp.list} (gplist, head=30, scales=list(x = list(rot =
    90)), highlight, ...)}
\description{method name}
\details{method description}
\arguments{\item{gplist}{...}
\item{head}{...}
\item{scales}{...}
\item{highlight}{...}
\item{...}{Further arguments passed to method}}
