\name{auroc}
\alias{auroc}
\title{Compute the area under an ROC curve.}
\usage{auroc(x, ...)}

\description{
  Compute the area under an ROC curve.
}

\details{
  Generic function for computing the area under an ROC
  curve.
}
\seealso{\code{\link{auroc.ep}}, \code{\link{auroc.ep.list}}}
\value{The area under the ROC curve(s).}
\arguments{
  \item{x}{An appropriate object.}
  \item{...}{Further arguments passed to method.}
}
