\name{bayes}
\alias{bayes}
\title{Bayesian posterior parameter estimates.}
\usage{bayes(x, ...)}
\description{Bayesian posterior parameter estimates.}
\details{A generic}
\seealso{\code{\link{bayes.bn}}, \code{\link{ml}}}
\arguments{\item{x}{An object}
\item{...}{Further arguments passed to method}}
