\name{bf.bnpostmcmc}
\alias{bf.bnpostmcmc}
\title{Bayes Factors.}
\usage{\method{bf}{bnpostmcmc} (bn1, bn2, data, ...)}

\description{
  Bayes Factors.
}

\details{
  method description
}
\seealso{\code{\link{bf}}}
\arguments{
  \item{bn1}{A bn}
  \item{bn2}{...}
  \item{data}{...}
  \item{...}{further arguments}
}
