\name{bf}
\alias{bf}
\title{Bayes Factor.}
\usage{bf(bn1, bn2, data, ...)}

\description{
  Bayes Factor.
}

\details{
  method description
}
\seealso{\code{\link{bf.bnpostmcmc}}}
\arguments{
  \item{bn1}{A bn}
  \item{bn2}{A bn}
  \item{data}{data}
  \item{...}{Further arguments passed to method}
}
