% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-mcmc.R
\name{bnpostmcmc.list}
\alias{bnpostmcmc.list}
\title{List of BN Posteriors from MCMC.}
\usage{
bnpostmcmc.list(...)
}
\arguments{
\item{...}{Further arguments passed to method}
}
\description{
method description
}
\seealso{
\code{\link{bnpostmcmc}}, \code{\link{gp.bnpostmcmc.list}},
  \code{\link{ep.bnpostmcmc.list}}
}
