\name{bnpostmcmc.list}
\alias{bnpostmcmc.list}
\title{List of BN Posteriors from MCMC.}
\usage{\method{bnpostmcmc}{list} (...)}
\description{List of BN Posteriors from MCMC.}
\details{method description}
\seealso{\code{\link{bnpostmcmc}}, \code{\link{gp.bnpostmcmc.list}},
\code{\link{ep.bnpostmcmc.list}}}
\arguments{\item{...}{Further arguments passed to method}}
