\name{BNSamplerPT}
\alias{BNSamplerPT}
\title{method name...}
\usage{BNSamplerPT(data, initial, prior, return="network",
    logScoreFUN=defaultLogScoreFUN(),
    logScoreParameters=list(hyperparameters = "qi"), constraint,
    modejumping=F, verbose=F, keepTape=F, temp=1, cache=new.env(hash =
    T, size = 10000L))}
\description{method name}
\details{method description}
\arguments{\item{data}{...}
\item{initial}{...}
\item{prior}{...}
\item{return}{...}
\item{logScoreFUN}{...}
\item{logScoreParameters}{...}
\item{constraint}{...}
\item{modejumping}{...}
\item{verbose}{...}
\item{keepTape}{...}
\item{temp}{...}
\item{cache}{...}}
