\name{burnin}
\alias{burnin}
\title{Extract burnin...}
\usage{burnin(x, ...)}

\description{
  Extract burnin
}

\details{
  Extracts the amount of burnin used in a sampler
}
\value{An integer amount of burnin}
\seealso{\code{\link{steps.sampler}}}
\arguments{
  \item{x}{A sampler}
  \item{...}{Further arguments passed to method}
}
