\name{burnin.sampler}
\alias{burnin.sampler}
\title{Extract burnin...}
\usage{\method{burnin}{sampler} (x, ...)}

\description{
  Extract burnin
}

\details{
  Extracts the amount of burnin used in a sampler
}
\value{An integer amount of burnin}
\arguments{
  \item{x}{A sampler}
  \item{...}{Further arguments, currently unused}
}
