\name{checkForNonDuplicated}
\alias{checkForNonDuplicated}
\title{Remove options that don't duplicate required duplicates.}
\usage{checkForNonDuplicated(ll, duplicates)}
\description{Remove options that don't duplicate required duplicates.}
\value{A logical vector}
\arguments{\item{ll}{A list}
\item{duplicates}{A table giving the number of appearances of each}}
