\name{checkForNonDuplicated}
\alias{checkForNonDuplicated}
\title{Remove options that don't duplicate required duplicates.}
\usage{
  checkForNonDuplicated(ll, duplicates)
}
\arguments{
  \item{ll}{A list}

  \item{duplicates}{A table giving the number of
  appearances of each}
}
\value{
  A logical vector
}
\description{
  Remove options that don't duplicate required duplicates.
}

