\name{cumep.bnpostmcmc.list}
\alias{cumep.bnpostmcmc.list}
\title{Cumulative edge probabilities.}
\usage{\method{cumep}{bnpostmcmc.list} (x, ...)}

\description{
  Cumulative edge probabilities.
}

\details{
  method description
}
\seealso{\code{\link{cumep}}}
\arguments{
  \item{x}{...}
  \item{...}{Further arguments passed to method}
}
