\name{cumep.samplers}
\alias{cumep.samplers}
\title{Cumulative edge probabilities.}
\usage{\method{cumep}{samplers} (x, ...)}

\description{
  Cumulative edge probabilities.
}

\details{
  method description
}
\seealso{\code{\link{cumep}}}
\arguments{
  \item{x}{...}
  \item{...}{Further arguments passed to method}
}
