% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{cummean.matrix}
\alias{cummean.matrix}
\title{Cumulative mean.}
\usage{
\method{cummean}{matrix}(x, ...)
}
\arguments{
\item{x}{A matrix}

\item{...}{Further arguments (unused)}
}
\value{
A matrix of the same dimension as x, with the cumulative values.
}
\description{
Compute the cumulative means of the columns of a matrix x.
ie each column is treated separately
}
\seealso{
\code{\link{cummean}}
}
