\name{cummean.matrix}
\alias{cummean.matrix}
\title{Cumulative mean.}
\usage{\method{cummean}{matrix} (x, ...)}

\description{
  Cumulative mean.
}

\details{
  Compute the cumulative means of the columns of a matrix
  x. ie each column is treated separately
}
\value{A matrix of the same dimension as x, with the cumulative values.}
\seealso{\code{\link{cummean}}}
\arguments{
  \item{x}{A matrix}
  \item{...}{Further arguments (unused)}
}
