\name{cumtvd}
\alias{cumtvd}
\title{method name...}
\usage{cumtvd(exactgp, bnpostmcmclist, start=1, end, nbin=floor((start -
    end)/100))}
\description{method name}
\details{method description}
\arguments{\item{exactgp}{...}
\item{bnpostmcmclist}{...}
\item{start}{...}
\item{end}{...}
\item{nbin}{...}
\item{...}{Further arguments passed to method}}
