\name{cumtvd}
\alias{cumtvd}
\title{Cumulative total variation distance.}
\usage{cumtvd(exactgp, bnpostmcmclist, start=1, end, nbin=floor((start -
    end)/100))}
\description{Cumulative total variation distance.}
\details{method description}
\seealso{\code{\link{xyplot.cumtvd}}, \code{\link{cumtvd.list}}}
\arguments{\item{exactgp}{...}
\item{bnpostmcmclist}{...}
\item{start}{...}
\item{end}{...}
\item{nbin}{...}
\item{...}{Further arguments passed to method}}
