\name{dagcv}
\alias{dagcv}
\title{method name...}
\usage{dagcv(graph, response, data, fold=10, verbose=F, binomial=F,
    predictiveReplications=10000, Freq, ...)}
\description{method name}
\details{method description}
\arguments{\item{graph}{...}
\item{response}{...}
\item{data}{...}
\item{fold}{...}
\item{verbose}{...}
\item{binomial}{...}
\item{predictiveReplications}{...}
\item{...}{Further arguments passed to method}
\item{Freq}{This argument was missing, and was picked up by R-check}}
