\name{defaultBanned}
\alias{defaultBanned}
\title{Maximal banned list.}
\usage{defaultBanned(nodesSeq)}
\description{Maximal banned list.}
\value{A list, which each element containing a numeric vector of the
nodes that are banned from being parents of the corresponding node.}
\arguments{\item{nodesSeq}{A sequence \code{1:numberOfNodes}}}
