\name{dotplot.ep.list}
\alias{dotplot.ep.list}
\title{Dotplot of list of posterior edge probabilities.}
\usage{\method{dotplot}{ep.list} (eplist, subsetBy="Exact", head=30, ...)}
\description{Dotplot of list of posterior edge probabilities.}
\details{method description}
\seealso{\code{\link{dotplot.ep}}}
\arguments{\item{eplist}{...}
\item{subsetBy}{...}
\item{head}{...}
\item{...}{Further arguments passed to method}}
