\name{dotplot.ep}
\alias{dotplot.ep}
\title{Dotplot of posterior edge probabilities.}
\usage{\method{dotplot}{ep} (ep, head=30, ...)}
\description{Dotplot of posterior edge probabilities.}
\details{method description}
\arguments{\item{ep}{...}
\item{head}{...}
\item{...}{Further arguments passed to method}}
