\name{dotplot.gp.list}
\alias{dotplot.gp.list}
\title{method name...}
\usage{\method{dotplot}{gp.list} (gplist, subsetBy="Exact", head=30, ...)}
\description{method name}
\details{method description}
\arguments{\item{gplist}{...}
\item{subsetBy}{...}
\item{head}{...}
\item{...}{Further arguments passed to method}}
