\name{drawPT}
\alias{drawPT}
\title{method name...}
\usage{drawPT(samplers, data, n, temperatures, pswap=0.2, burnin=0, verbose=T,
    currentNetwork)}
\description{method name}
\details{method description}
\arguments{\item{samplers}{...}
\item{data}{...}
\item{n}{...}
\item{temperatures}{...}
\item{pswap}{...}
\item{burnin}{...}
\item{verbose}{...}
\item{currentNetwork}{BUG found by R-check}}
