% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{enforceForbiddenConstraint}
\alias{enforceForbiddenConstraint}
\title{Undocumented.}
\usage{
enforceForbiddenConstraint(x, constraint)
}
\arguments{
\item{x}{...}

\item{constraint}{...}
}
\description{
method description
}
\seealso{
\code{\link{satisfiesConstraint}}, \code{\link{setupConstraint}}
}
