\name{enforceForbiddenConstraint}
\alias{enforceForbiddenConstraint}
\title{Undocumented.}
\usage{enforceForbiddenConstraint(x, constraint)}

\description{
  Undocumented.
}

\details{
  method description
}
\seealso{\code{\link{satisfiesConstraint}}, \code{\link{setupConstraint}}}
\arguments{
  \item{x}{...}
  \item{constraint}{...}
}
