\name{ep.bnpostmcmc.list}
\alias{ep.bnpostmcmc.list}
\title{Posterior edge probabilities.}
\usage{\method{ep}{bnpostmcmc.list} (x, start, end, ...)}
\description{Posterior edge probabilities.}
\details{method description}
\seealso{\code{\link{ep}}, \code{\link{ep.bnpostmcmc}}}
\arguments{\item{x}{...}
\item{start}{...}
\item{end}{...}
\item{...}{further arguments}}
