\name{epmx}
\alias{epmx}
\title{Edge probabilties matrix.}
\usage{epmx(x, ...)}
\description{Edge probabilties matrix.}
\details{method description}
\seealso{\code{\link{epmx.bnpostmcmc.list}}, \code{\link{xyplot.epmx}},
\code{\link{splom.bnpostmcmc.list}}}
\arguments{\item{x}{...}
\item{...}{Further arguments passed to method}}
