\name{getNijkCounts}
\alias{getNijkCounts}
\title{Undocumented.}
\usage{getNijkCounts(data, bn, nodes=seq_along(bn), cache=new.env(hash = T))}
\description{Undocumented.}
\details{method description}
\seealso{\code{\link{ml}}, \code{\link{bayes}}}
\arguments{\item{data}{...}
\item{bn}{...}
\item{nodes}{...}
\item{cache}{...}}
