\name{gp.bnpost}
\alias{gp.bnpost}
\title{Posterior graph probaiblities.}
\usage{\method{gp}{bnpost} (x, logNetworkPriors, log=F, pretty=F, ...)}
\description{Posterior graph probaiblities.}
\details{method description}
\seealso{\code{\link{gp}}, \code{\link{ep.bnpost}}}
\arguments{\item{x}{...}
\item{logNetworkPriors}{...}
\item{log}{...}
\item{pretty}{...}
\item{...}{Further arguments (unused)}}
