\name{gp.bnpostmcmc.list}
\alias{gp.bnpostmcmc.list}
\title{Posterior graph probabilities.}
\usage{\method{gp}{bnpostmcmc.list} (x, ...)}
\description{Posterior graph probabilities.}
\details{method description}
\seealso{\code{\link{gp}}, \code{\link{gp.bnpostmcmc}}}
\arguments{\item{x}{...}
\item{...}{Further arguments passed to method}}
