\name{gp.bnpostmcmc}
\alias{gp.bnpostmcmc}
\title{Posterior graph probabilities.}
\usage{
  \method{gp}{bnpostmcmc} (x, start, end, nbin = 1, log =
  F, pretty = F, levels = NULL, ...)
}
\arguments{
  \item{x}{...}

  \item{start}{...}

  \item{end}{...}

  \item{nbin}{...}

  \item{log}{...}

  \item{pretty}{...}

  \item{levels}{...}

  \item{...}{Further arguments (unused)}
}
\description{
  method description
}
\seealso{
  \code{\link{gp}}, \code{\link{gp.bnpostmcmc.list}}
}

