\name{gp.bnpostmcmc}
\alias{gp.bnpostmcmc}
\title{Posterior graph probabilities.}
\usage{\method{gp}{bnpostmcmc} (x, logNetworkPriors, start, end, nbin=1,
    log=F, pretty=F, levels, ...)}
\description{Posterior graph probabilities.}
\details{method description}
\arguments{\item{x}{...}
\item{logNetworkPriors}{...}
\item{start}{...}
\item{end}{...}
\item{nbin}{...}
\item{log}{...}
\item{pretty}{...}
\item{levels}{...}
\item{...}{Further arguments (unused)}}
