\name{gp}
\alias{gp}
\title{Posterior graph probabilities.}
\usage{gp(x, ...)}

\description{
  Posterior graph probabilities.
}

\details{
  method description
}
\seealso{\code{\link{summary.gp}}, 
\code{\link{gp.bnpostmcmc}}, \code{\link{gp.bnpost}}, 
\code{\link{gp.bnpostmcmc.list}}, \code{\link{gp.list}}}
\arguments{
  \item{x}{...}
  \item{...}{Further arguments passed to method}
}
