\name{intAsFDF}
\alias{intAsFDF}
\title{Convert a data frame of integers to factors.}
\usage{intAsFDF(x)}

\description{
  Convert a data frame of integers to factors.
}

\details{
  Converts a \code{data.frame} that consists columns of
  integers into a \code{data.frame} consisting of factos
}
\value{The data.frame with columns converted to factors}
\seealso{\code{\link{fdfAsInt}}}
\arguments{
  \item{x}{an object of class \code{data.frame}}
}
