\name{intAsFDF}
\alias{intAsFDF}
\title{Convert data...}
\usage{intAsFDF(x)}
\description{Convert \code{data.frame} of integers to factors}
\details{Converts a \code{data.frame} that consists columns of integers into a
\code{data.frame} consisting of factos}
\value{The data.frame with columns converted to factors}
\arguments{\item{x}{an object of class \code{data.frame}}}
