\name{is.valid.hyp}
\alias{is.valid.hyp}
\title{Checks validity.}
\usage{
  \method{is.valid}{hyp} (x)
}
\arguments{
  \item{x}{The hyperparameters to test.}
}
\value{
  A logical of length 1.
}
\description{
  Checks whether the supplied hyperparameters parameter is
  valid.
}
\seealso{
  \code{\link{logScoreMultDir}}
}

