\name{is.valid.localPriors}
\alias{is.valid.localPriors}
\title{Check validity.}
\usage{
  is.valid.localPriors(x)
}
\arguments{
  \item{x}{An R object}
}
\value{
  A logical of length 1, indicating the validity of x
}
\description{
  Checks whether the supplied R object is a valid
  \code{localPriors}
}

