\name{is.valid.localPriors}
\alias{is.valid.localPriors}
\title{Check validity.}
\usage{is.valid.localPriors(x)}

\description{
  Check validity.
}

\details{
  Checks whether the supplied R object is a valid
  \code{localPriors}
}
\value{A logical of length 1, indicating the validity of x}
\arguments{
  \item{x}{An R object}
}
