\name{length.bnpostmcmc}
\alias{length.bnpostmcmc}
\title{Returns the number of samples draw in the supplied object of class...}
\usage{\method{length}{bnpostmcmc} (x, ...)}
\description{Returns the number of samples draw in the supplied object of class
'bnpostmcmc' x.}
\value{The number of samples in x}
\arguments{\item{x}{An object of class 'bnpostmcmc'}
\item{...}{Further arguments passed to method}}
