\name{length.bnpostmcmc}
\alias{length.bnpostmcmc}
\title{Number of samples drawn.}
\usage{\method{length}{bnpostmcmc} (x, ...)}

\description{
  Number of samples drawn.
}

\details{
  Returns the number of samples draw in the supplied object
  of class 'bnpostmcmc' x.
}
\value{The number of samples in x}
\seealso{\code{\link{bnpostmcmc}}.}
\arguments{
  \item{x}{An object of class 'bnpostmcmc'}
  \item{...}{Further arguments passed to method}
}
