% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{levelplot.bnpostmcmc}
\alias{levelplot.bnpostmcmc}
\title{Levelplot of BN Posterior from MCMC.}
\usage{
\method{levelplot}{bnpostmcmc}(x)
}
\arguments{
\item{x}{...}
}
\description{
method description
}
\seealso{
\code{\link{plot.bnpostmcmc}}
}
