\name{localPartitionFunction}
\alias{localPartitionFunction}
\title{Undocumented.}
\usage{
  localPartitionFunction(currentNetwork, proposalNetwork,
    node, required = integer(0), logScoreOnlineFUN,
    logScoreParameters, cache)
}
\arguments{
  \item{currentNetwork}{...}

  \item{proposalNetwork}{...}

  \item{node}{...}

  \item{required}{...}

  \item{logScoreOnlineFUN}{...}

  \item{logScoreParameters}{...}

  \item{cache}{...}
}
\description{
  method description
}

