\name{map}
\alias{map}
\title{Maximum aposteriori graph.}
\usage{map(x, ...)}

\description{
  Maximum aposteriori graph.
}

\details{
  method description
}
\seealso{\code{\link{map.bnpost}}, \code{\link{map.bnpostmcmc}}}
\arguments{
  \item{x}{...}
  \item{...}{Further arguments passed to method}
}
