\name{mwep.bnpostmcmc.list}
\alias{mwep.bnpostmcmc.list}
\title{Moving window edge probabilities.}
\usage{\method{mwep}{bnpostmcmc.list} (x, window=10, ...)}

\description{
  Moving window edge probabilities.
}

\details{
  method description
}
\seealso{\code{\link{mwep}}}
\arguments{
  \item{x}{...}
  \item{window}{...}
  \item{...}{Further arguments passed to method}
}
