\name{mwep.samplers}
\alias{mwep.samplers}
\title{Moving window edge probabilities.}
\usage{\method{mwep}{samplers} (x, window=10, ...)}

\description{
  Moving window edge probabilities.
}

\details{
  method description
}
\seealso{\code{\link{mwep}}}
\arguments{
  \item{x}{...}
  \item{window}{...}
  \item{...}{Further arguments passed to method}
}
