\name{mwmean.matrix}
\alias{mwmean.matrix}
\title{Compute the moving window means of the columns of a matrix x.}
\usage{\method{mwmean}{matrix} (x, window, ...)}
\description{Compute the moving window means of the columns of a matrix x.
ie each column is treated separately}
\value{A matrix of the same dimension as x, with the moving window values.}
\arguments{\item{x}{A matrix}
\item{window}{...}
\item{...}{Further arguments (unused)}}
