\name{rowSums2}
\alias{rowSums2}
\title{Fast, dangerous row sums.}
\usage{rowSums2(x)}
\description{Fast, dangerous row sums.}
\details{A fast, simple version of \code{rowSums}.
This version only handles matrices.

Note that no sanity checks are performed on the input.}
\value{A vector of row sums.}
\arguments{\item{x}{A matrix.}}
