\name{satisfiesConstraint}
\alias{satisfiesConstraint}
\title{Undocumented.}
\usage{
  satisfiesConstraint(x, constraint)
}
\arguments{
  \item{x}{...}

  \item{constraint}{...}
}
\description{
  method description
}
\seealso{
  \code{\link{setupConstraint}},
  \code{\link{enforceForbiddenConstraint}}
}

