\name{scoreBIC}
\alias{scoreBIC}
\title{Undocumented.}
\usage{scoreBIC(bn, data, nodes=seq_along(bn), mlcache=new.env(hash = T),
    nijkcache=new.env(hash = T))}
\description{Undocumented.}
\details{method description}
\arguments{\item{bn}{...}
\item{data}{...}
\item{nodes}{...}
\item{mlcache}{...}
\item{nijkcache}{...}}
