% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-bic.R
\name{scoreBIC}
\alias{scoreBIC}
\title{Undocumented.}
\usage{
scoreBIC(
  bn,
  data,
  nodes = seq_along(bn),
  mlcache = new.env(hash = T),
  nijkcache = new.env(hash = T)
)
}
\arguments{
\item{bn}{...}

\item{data}{...}

\item{nodes}{...}

\item{mlcache}{...}

\item{nijkcache}{...}
}
\description{
method description
}
