\name{scoreParents}
\alias{scoreParents}
\title{Undocumented.}
\usage{scoreParents(possibleParents, head, proposal, current,
    logScoreOnlineFUN, logScoreParameters, cache)}
\description{Undocumented.}
\details{method description}
\arguments{\item{possibleParents}{...}
\item{head}{...}
\item{proposal}{...}
\item{current}{...}
\item{logScoreOnlineFUN}{...}
\item{logScoreParameters}{...}
\item{cache}{...}}
