% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-grzeg.R
\name{scoreParents}
\alias{scoreParents}
\title{Undocumented.}
\usage{
scoreParents(
  possibleParents,
  head,
  proposal,
  current,
  logScoreOnlineFUN,
  logScoreParameters,
  cache
)
}
\arguments{
\item{possibleParents}{...}

\item{head}{...}

\item{proposal}{...}

\item{current}{...}

\item{logScoreOnlineFUN}{...}

\item{logScoreParameters}{...}

\item{cache}{...}
}
\description{
method description
}
