\name{setupConstraint}
\alias{setupConstraint}
\title{Setup constraint.}
\usage{setupConstraint(constraint, initial)}
\description{Setup constraint.}
\details{...}
\seealso{\code{\link{satisfiesConstraint}},
\code{\link{enforceForbiddenConstraint}}}
\arguments{\item{constraint}{..}
\item{initial}{..}}
