\name{splom.epmx}
\alias{splom.epmx}
\title{Plot of cumulative edge probabilities.}
\usage{\method{splom}{epmx} (x, subset)}

\description{
  Plot of cumulative edge probabilities.
}

\details{
  Returns a splom of the cumulative edge probabilities
  through time for bnpostmcmc.list and bvspostmcmc.list
}
\value{...}
\seealso{\code{\link{epmx}}}
\arguments{
  \item{x}{...}
  \item{subset}{A numeric vector. This specifies the nodes between which
the epmx should be displayed for.}
}
