\name{statistics}
\alias{statistics}
\title{Extract statistics from a sampler.}
\usage{
  statistics(x, ...)
}
\arguments{
  \item{x}{A sampler}

  \item{...}{Further arguments passed to method}
}
\value{
  A list of statistics collected during the MCMC run.
}
\description{
  Extracts the statistics collected during an MCMC run
}
\seealso{
  \code{\link{statistics.sampler}}
}

