% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{statistics.sampler}
\alias{statistics.sampler}
\title{Extract statistics from a sampler.}
\usage{
\method{statistics}{sampler}(x, names, ...)
}
\arguments{
\item{x}{A sampler}

\item{names}{Which statistics to extract? A character vector of
statistics collected during the MCMC run.}

\item{...}{Further arguments passed to method}
}
\value{
A list of statistics collected during the MCMC run.
}
\description{
Extracts the statistics collected during an MCMC run
}
\seealso{
\code{\link{statistics}}
}
