\name{summary.sampler}
\alias{summary.sampler}
\title{Summarising MCMC samplers...}
\usage{\method{summary}{sampler} (x, ...)}

\description{
  Summarising MCMC samplers
}

\details{
  \code{summary} method for class "\code{sampler}"
}
\arguments{
  \item{x}{A sampler}
  \item{...}{Further arguments, currently unused}
}
