\name{top.bnpost}
\alias{top.bnpost}
\title{Top graph from BN Posterior.}
\usage{\method{top}{bnpost} (x, head=10, ...)}
\description{Top graph from BN Posterior.}
\details{method description}
\seealso{\code{\link{top}}, \code{\link{map.bnpost}}}
\arguments{\item{x}{...}
\item{head}{...}
\item{...}{Further arguments (unused)}}
