\name{top.sampler}
\alias{top.sampler}
\title{Retreive top graph from sampler}
\usage{
  \method{top}{sampler} (x, head = 10, ...)
}
\arguments{
  \item{x}{A sampler}

  \item{head}{The number of graphs to consider}

  \item{...}{Further arguments, currently unused}
}
\description{
  \code{top} method for class "\code{sampler}"
}

