\name{whichNum}
\alias{whichNum}
\title{Find rows satisfying equality.}
\usage{whichNum(col, x, i)}
\description{Find rows satisfying equality.}
\details{Find which rows in column \code{col} of a matrix \code{x} are equal to 
\code{i}.

Note that no sanity checks are performed on the input.}
\value{A numeric vector, of the}
\arguments{\item{col}{The column of the matrix to use. A numeric vector of length 1.}
\item{x}{A matrix.}
\item{i}{The value to check for}}
