% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{xyplot.cumtvd.list}
\alias{xyplot.cumtvd.list}
\title{Plot of cumulative total variation distance.}
\usage{
\method{xyplot}{cumtvd.list}(cumtvdlist)
}
\arguments{
\item{cumtvdlist}{...}
}
\description{
method description
}
\seealso{
\code{\link{xyplot.cumtvd.list}}
}
