\name{xyplot.epmx}
\alias{xyplot.epmx}
\title{Returns a xyplot of the cumulative edge probabilities through time...}
\usage{\method{xyplot}{epmx} (x)}
\description{Returns a xyplot of the cumulative edge probabilities through time
for bnpostmcmc.list and bvspostmcmc.list}
\value{...}
\arguments{\item{x}{...}}
