\name{xyplot.epmx}
\alias{xyplot.epmx}
\title{Plot of cumulative edge probabilities.}
\usage{\method{xyplot}{epmx} (x)}
\description{Plot of cumulative edge probabilities.}
\details{Returns a xyplot of the cumulative edge probabilities through time
for bnpostmcmc.list and bvspostmcmc.list}
\value{...}
\seealso{\code{\link{epmx}}}
\arguments{\item{x}{...}}
