\name{xyplot.gp.list}
\alias{xyplot.gp.list}
\title{Scatterplot of posterior graph probablities.}
\usage{\method{xyplot}{gp.list} (gplist, head=30, scales=list(x = list(rot =
    90)), highlight, ...)}
\description{Scatterplot of posterior graph probablities.}
\details{method description}
\seealso{\code{\link{dotplot.gp.list}}, \code{\link{xyplot.gp}}}
\arguments{\item{gplist}{...}
\item{head}{...}
\item{scales}{...}
\item{highlight}{...}
\item{...}{Further arguments passed to method}}
