\name{xyplot.gp}
\alias{xyplot.gp}
\title{Scatterplot of posterior graph probabilities.}
\usage{\method{xyplot}{gp} (gp, head=30, ...)}
\description{Scatterplot of posterior graph probabilities.}
\details{method description}
\arguments{\item{gp}{...}
\item{head}{...}
\item{...}{Further arguments passed to method}}
